# src/swagger.json
{
	"components": {
...
		"schemas": {
			"ProfileType": {
				"type": "string",
				"enum": [
					"stakeholder",
					"admin",
					"unit owner"
				]
			},
			"RecordStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive"
				]
			},
...
			"CommunityAttr": {
				"properties": {
					"description": {
						"type": "string"
					}
				},
				"required": [
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChargeType": {
				"type": "string",
				"enum": [
					"unit",
					"percentage",
					"amount"
				]
			},
			"PostingType": {
				"type": "string",
				"enum": [
					"monthly",
					"manual",
					"accrued",
					"interest",
					"quarterly"
				]
			},
			"ChargeAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"communityId": {
						"type": "number",
						"format": "double"
					},
					"community": {
						"$ref": "#/components/schemas/CommunityAttr"
					},
					"code": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"chargeType": {
						"$ref": "#/components/schemas/ChargeType"
					},
					"postingType": {
						"$ref": "#/components/schemas/PostingType"
					},
					"thresholdInMonths": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"priority": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"passOn": {
						"type": "boolean",
						"nullable": true
					}
				},
				"required": [
					"communityId",
					"code",
					"rate",
					"chargeType",
					"postingType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaymentType": {
				"type": "string",
				"enum": [
					"cash",
					"check"
				]
			},
			"ProfileAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"scopes": {
						"type": "string"
					},
					"remarks": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"email",
					"type",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DisbursementAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"releasedBy": {
						"type": "number",
						"format": "double"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"details": {
						"type": "string"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"releasedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"chargeId",
					"releasedBy",
					"paymentType",
					"details",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RequestStatus": {
				"type": "string",
				"enum": [
					"approved",
					"rejected",
					"pending",
					"cancelled"
				]
			},
			"ExpenseAttr": {
				"properties": {
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitCost": {
						"type": "number",
						"format": "double"
					},
					"totalCost": {
						"type": "number",
						"format": "double"
					},
					"waivedBy": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"categoryId",
					"category",
					"description",
					"quantity",
					"unitCost",
					"totalCost"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreatePurchaseOrder": {
				"properties": {
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"otherDetails": {
						"type": "string"
					},
					"fulfillmentDate": {
						"type": "string",
						"format": "date-time"
					},
					"vendorName": {
						"type": "string"
					}
				},
				"required": [
					"purchaseRequestId",
					"otherDetails",
					"fulfillmentDate",
					"vendorName"
				],
				"type": "object"
			},
			"CreateVoucherOrOrder": {
				"properties": {
					"orderData": {
						"$ref": "#/components/schemas/CreatePurchaseOrder"
					},
					"expenses": {
						"items": {
							"$ref": "#/components/schemas/ExpenseAttr"
						},
						"type": "array"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"requestedDate": {
						"type": "string",
						"format": "date-time"
					},
					"requestedBy": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"expenses",
					"chargeId",
					"requestedDate",
					"requestedBy",
					"description"
				],
				"type": "object"
			},
			"ApproveVoucherOrOrder": {
				"properties": {
					"disbursements": {
						"items": {
							"$ref": "#/components/schemas/DisbursementAttr"
						},
						"type": "array"
					},
					"codes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"codes"
				],
				"type": "object"
			},
...
			"VoucherAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"totalCost": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/RequestStatus"
					},
					"requestedBy": {
						"type": "number",
						"format": "double"
					},
					"requestedDate": {
						"type": "string",
						"format": "date-time"
					},
					"approvedBy": {
						"type": "string",
						"nullable": true
					},
					"rejectedBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string",
						"nullable": true
					},
					"expenses": {
						"items": {
							"$ref": "#/components/schemas/ExpenseAttr"
						},
						"type": "array"
					},
					"disbursements": {
						"items": {
							"$ref": "#/components/schemas/DisbursementAttr"
						},
						"type": "array"
					},
					"requestedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"rejectedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"approverProfiles": {
						"items": {
							"$ref": "#/components/schemas/ProfileAttr"
						},
						"type": "array"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"series": {
						"type": "string"
					}
				},
				"required": [
					"chargeId",
					"description",
					"totalCost",
					"status",
					"requestedBy",
					"requestedDate",
					"series"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RejectVoucher": {
				"properties": {
					"rejectedBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rejectedBy",
					"comments",
					"id"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "authorization"
			}
		}
	},
	"info": {
		"title": "luminous-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/voucher/getVoucher/{id}": {
			"get": {
				"operationId": "GetVoucher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VoucherAttr"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/voucher/getAllVouchersByChargeAndStatus/{chargeId}/{status}": {
			"get": {
				"operationId": "GetAllVouchersByChargeAndStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/VoucherAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chargeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RequestStatus"
						}
					}
				]
			}
		},
		"/api/voucher/postVoucher": {
			"post": {
				"operationId": "PostVoucher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/voucher/updateVoucher/{id}": {
			"patch": {
				"operationId": "UpdateVoucher",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/voucher/approveVoucher": {
			"post": {
				"operationId": "ApproveVoucher",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApproveVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/voucher/rejectVoucher": {
			"post": {
				"operationId": "RejectVoucher",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RejectVoucher"
							}
						}
					}
				}
			}
		},
		"/api/voucher/notifyVoucherApprovers/{id}": {
			"post": {
				"operationId": "NotifyVoucherApprovers",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}